repo reorg: move health check filter
