+++ source/extensions/extensions_build_config.bzl
@@ -13,6 +13,7 @@ EXTENSIONS = {
 
     "envoy.filters.http.dynamo":                "//source/extensions/filters/http/dynamo:config",
     "envoy.filters.http.ext_authz":             "//source/extensions/filters/http/ext_authz:config",
+    "envoy.filters.http.health_check":          "//source/extensions/filters/http/health_check:config",
     "envoy.filters.http.lua":                   "//source/extensions/filters/http/lua:config",
     "envoy.filters.http.ratelimit":             "//source/extensions/filters/http/ratelimit:config",
