+++ source/server/http/BUILD
@@ -65,27 +65,3 @@ envoy_cc_library(
         "//source/common/common:macros",
     ],
 )
-
-envoy_cc_library(
-    name = "health_check_lib",
-    srcs = ["health_check.cc"],
-    hdrs = ["health_check.h"],
-    deps = [
-        "//include/envoy/event:dispatcher_interface",
-        "//include/envoy/event:timer_interface",
-        "//include/envoy/http:codes_interface",
-        "//include/envoy/http:filter_interface",
-        "//include/envoy/http:header_map_interface",
-        "//include/envoy/server:filter_config_interface",
-        "//source/common/common:assert_lib",
-        "//source/common/common:enum_to_int",
-        "//source/common/config:filter_json_lib",
-        "//source/common/config:well_known_names",
-        "//source/common/http:codes_lib",
-        "//source/common/http:header_map_lib",
-        "//source/common/http:headers_lib",
-        "//source/common/http:utility_lib",
-        "//source/common/protobuf:utility_lib",
-        "@envoy_api//envoy/config/filter/http/health_check/v2:health_check_cc",
-    ],
-)
