+++ test/server/BUILD
@@ -156,6 +156,7 @@ envoy_cc_test(
         "//source/common/common:version_lib",
         "//source/extensions/access_loggers/file:config",
         "//source/extensions/filters/http/dynamo:config",
+        "//source/extensions/filters/http/health_check:config",
         "//source/extensions/filters/http/ratelimit:config",
         "//source/extensions/filters/network/hcm:config",
         "//source/extensions/filters/network/redis_proxy:config",
